requirements.txt
cryptography
pillow
numpy
pycryptodome
jpegio